Separate Table Panel from the Report
