+++ src/views/ADempiere/ReportViewerEngine/reportPanel.vue
@@ -1,7 +1,7 @@
 <!--
   ADempiere-Vue (Frontend) for ADempiere ERP & CRM Smart Business Solution
   Copyright (C) 2018-Present E.R.P. Consultores y Asociados, C.A. www.erpya.com
-  Contributor(s): Edwin Betancourt EdwinBetanc0urt@outlook.com https://github.com/EdwinBetanc0urt
+  Contributor(s): Elsio Sanchez elsiosanches@gmail.com https://github.com/ElsioSanchez
   This program is free software: you can redistribute it and/or modify
   it under the terms of the GNU General Public License as published by
   the Free Software Foundation, either version 3 of the License, or
@@ -16,99 +16,51 @@
   along with this program. If not, see <https:www.gnu.org/licenses/>.
 -->
 <template>
-  <div @click="showPopover = false">
-    <el-card class="containerReportEnginer">
-      <reportSearchCriteria
-        :container-uuid="reportOutput.containerUuid"
+  <el-card class="containerReportEnginer">
+    <reportSearchCriteria
+      :container-uuid="reportOutput.containerUuid"
+      :report-output="reportOutput"
+    />
+    <el-dialog
+      v-shortkey="shortsKey"
+      :visible.sync="showDialog"
+      :title="$t('report.reportEnginer.optionsImport.title')"
+      top="2vh"
+      @shortkey.native="keyAction"
+      @close="viewShowDialog"
+    >
+      <dialogShareReport
         :report-output="reportOutput"
+        :container-uuid="containerUuid"
       />
-      <el-dialog
-        v-shortkey="shortsKey"
-        :visible.sync="showDialog"
-        :title="$t('report.reportEnginer.optionsImport.title')"
-        top="2vh"
-        @shortkey.native="keyAction"
-        @close="viewShowDialog"
-      >
-        <dialogShareReport
-          :report-output="reportOutput"
-          :container-uuid="containerUuid"
-        />
-      </el-dialog>
-      <el-table
-        ref="tableReportEngine"
-        v-loading="isLoadingReport"
-        :data="dataList"
-        row-key="level"
-        style="width: 100%;"
-        lazy
-        :row-class-name="tableRowClassName"
-        :default-expand-all="false"
-        :tree-props="{ children: 'children' }"
-        :height="height"
-        border
-        :cell-style="{ padding: '0', height: '30px', border: 'none' }"
-        :cell-class-name="getRowClassName"
-        @row-click="handleRowClick"
-        @cell-contextmenu="activatePopover"
-      >
-        <el-table-column
-          v-for="(fieldAttributes, key) in columns"
-          :key="fieldAttributes.code + key"
-          :column-key="fieldAttributes.code"
-          :align="getAlignment(fieldAttributes.display_type)"
-          :fixed="fieldAttributes.is_group_column"
-          :width="widthColumn()"
-        >
-          <template slot="header">
-            {{ fieldAttributes.title }}
-          </template>
-          <template slot-scope="scope">
-            <span
-              v-if="!shouldHideName(scope.row, fieldAttributes.code, key)"
-              :style="getCellStyle(fieldAttributes.code, scope.row)"
-            >
-              {{ displayLabel(fieldAttributes, scope.row) }}
-              <el-popover
-                v-if="selectedRow === scope.row && selectedColumn === fieldAttributes.code && scope.row.is_parent"
-                v-model="showPopover"
-                placement="top"
-                class="reportInfo"
-                style="position: fixed; z-index: 1000; background-color: #fff;"
-                @shortkey.native="keyAction"
-              >
-                <InfoReport
-                  :data="dataModal"
-                />
-              </el-popover>
-            </span>
-          </template>
-        </el-table-column>
-      </el-table>
-      <custom-pagination
-        :total-records="recordData.record_count"
-        :page-size="currentPageSize"
-        :page-number="currentPageNumber"
-        :handle-change-page-size="handleChangeSizePage"
-        :handle-change-page-number="handleChangePage"
-      />
-    </el-card>
-  </div>
+    </el-dialog>
+    <data-report
+      :container-manager="containerManager"
+      :instance-uuid="instanceUuid"
+      :container-uuid="containerUuid"
+      :report-output="reportOutput"
+    />
+  </el-card>
 </template>
 <script>
+import {
+  defineComponent,
+  computed
+} from '@vue/composition-api'
 import store from '@/store'
-import { defineComponent, computed, ref, watch, nextTick, onMounted } from '@vue/composition-api'
-import { isEmptyValue } from '@/utils/ADempiere/valueUtils.js'
+// Components
 import CustomPagination from '@/components/ADempiere/DataTable/Components/CustomPagination.vue'
-import { isNumberField, isLookup } from '@/utils/ADempiere/references'
-import InfoReport from './infoReport'
-import dialogShareReport from './dialog'
-import reportSearchCriteria from './searchCriteria'
+import InfoReport from '@/views/ADempiere/ReportViewerEngine/infoReport.vue'
+import dialogShareReport from '@/views/ADempiere/ReportViewerEngine/dialog'
+import reportSearchCriteria from '@/views/ADempiere/ReportViewerEngine/searchCriteria'
+import DataReport from '@/components/ADempiere/Report/DataReport.vue'
+
 export default defineComponent({
   name: 'reportPanel',
   components: {
     CustomPagination,
     InfoReport,
+    DataReport,
     dialogShareReport,
     reportSearchCriteria
   },
@@ -130,291 +82,36 @@ export default defineComponent({
       required: false
     }
   },
-  setup(props) {
-    const dataModal = ref({})
-    const showPopover = ref(false)
-    const selectedRow = ref(undefined)
-    const selectedColumn = ref(undefined)
-    const tableReportEngine = ref(undefined)
-    const height = computed(() => {
-      if (store.getters.device !== 'mobile') {
-        return 'calc(100vh - 275px)'
-      }
-      return 'calc(100vh - 385px)'
-    })
+  setup() {
+    // Computed
     const shortsKey = computed(() => {
       return {
         close: ['esc']
       }
     })
+
+    const showDialog = computed(() => {
+      return store.getters.getReportShowDialog
+    })
+    // Methods
     function keyAction(event) {
       switch (event.srcKey) {
         case 'close':
           viewShowDialog()
-          showPopover.value = false
       }
     }
-    const data = computed(() => {
-      const { rowCells } = props.reportOutput
-      if (isEmptyValue(rowCells)) return []
-      return rowCells
-    })
-    const columns = computed(() => {
-      const { columns } = props.reportOutput
-      if (isEmptyValue(columns)) return []
-      return columns
-    })
-    function handleRowClick(row) {
-      if (row.children && row.children.length > 0) {
-        tableReportEngine.value.toggleRowExpansion(row)
-        showPopover.value = false
-      }
-    }
-    function activatePopover(row, column) {
-      event.preventDefault()
-      Object.values(row.cells).forEach(dataCell => {
-        if (dataCell && 'sum_value' in dataCell) {
-          selectedColumn.value = column.columnKey
-          selectedRow.value = row
-          dataModal.value = dataCell
-          showPopover.value = true
-        }
-      })
-    }
-    function displayLabel(field, row) {
-      if (isEmptyValue(row.cells)) {
-        return
-      }
-      const rowData = row.cells[field.code]
-      if (!isEmptyValue(rowData)) {
-        const { display_value, value } = rowData
-        if (!isEmptyValue(display_value) || isLookup(field.display_type)) {
-          return display_value
-        }
-        if (!isEmptyValue(value)) {
-          return value
-        }
-      }
-    }
-    function getAlignment(displayType) {
-      if (isNumberField(displayType)) {
-        return 'right'
-      }
-      return 'left'
-    }
-    const dataList = computed(() => {
-      return data.value.map((row, rowIndex) => {
-        const index = rowIndex + 1
-        const newRow = {
-          ...row,
-          children: hasChildren(row.children, index.toString()),
-          level: index
-        }
-        return newRow
-      })
-    })
-    const recordData = computed(() => {
-      return store.getters.getReportOutput(props.instanceUuid)
-    })
-    const currentPageSize = computed(() => {
-      return parseInt(props.reportOutput.pageSize, 10)
-    })
-    const currentPageNumber = computed(() => {
-      return parseInt(props.reportOutput.pageToken, 10)
-    })
 
-    function hasChildren(children, parentLevel) {
-      if (children.length < 1) return children
-      return children.map((child, indexChild) => {
-        const index = parentLevel + indexChild
-        return {
-          ...child,
-          children: hasChildren(child.children, index.toString()),
-          level: index
-        }
-      })
-    }
     function viewShowDialog() {
       store.commit('setShowDialog', false)
     }
-    const showDialog = computed(() => {
-      return store.getters.getReportShowDialog
-    })
-    const reportDefinition = store.getters.getStoredReport(props.reportOutput.containerUuid)
-    function handleChangeSizePage(pageSize) {
-      props.containerManager.setPageSize({
-        instanceUuid: props.reportOutput.instance_id,
-        containerUuid: props.reportOutput.containerUuid,
-        pageNumber: currentPageNumber.value,
-        pageSize,
-        parametersList: reportDefinition,
-        reportId: reportDefinition.id,
-        printFormatId: props.reportOutput.print_format_id,
-        reportViewId: props.reportOutput.report_view_id
-      })
-    }
-    function handleChangePage(pageNumber) {
-      props.containerManager.setPageNumber({
-        instanceUuid: props.reportOutput.instance_id,
-        containerUuid: props.reportOutput.containerUuid,
-        pageNumber,
-        pageSize: currentPageSize.value,
-        parametersList: reportDefinition,
-        reportId: reportDefinition.id,
-        printFormatId: props.reportOutput.print_format_id,
-        reportViewId: props.reportOutput.report_view_id
-      })
-    }
 
-    function exportFile() {
-      store.dispatch('exportReport', {
-        reportId: props.reportOutput.id,
-        reportName: props.reportOutput.name
-      })
-    }
-    function tableRowClassName({ row, rowIndex }) {
-      const { children } = row
-      if (!isEmptyValue(children)) {
-        return 'success-row'
-      }
-      return ''
-    }
-    function getRowClassName({ row, rowIndex }) {
-      const parent = this.findParent(row)
-      if (parent && parent.children[parent.children.length - 1] === row) {
-        return 'last-child-row'
-      }
-      return ''
-    }
-    function findParent(row) {
-      let parentRow = null
-      const stack = [...this.dataList]
-      while (stack.length) {
-        const current = stack.pop()
-        if (current.children && current.children.includes(row)) {
-          parentRow = current
-          break
-        }
-        if (current.children) {
-          stack.push(...current.children)
-        }
-      }
-      return parentRow
-    }
-    const expanded = computed(() => {
-      return store.getters.getExpandedAll
-    })
-    function expandedRowAll() {
-      function expandRecursively(row) {
-        tableReportEngine.value.toggleRowExpansion(row, true)
-        if (row.children && row.children.length > 0) {
-          row.children.forEach(expandRecursively)
-        }
-      }
-      function collapseRecursively(row) {
-        let hideRow = false
-        if (row.children && row.children.length > 0) {
-          row.children.forEach(child => {
-            if (!child.is_parent) {
-              hideRow = true
-            }
-            collapseRecursively(child)
-          })
-        }
-        if (!row.is_parent || hideRow) {
-          tableReportEngine.value.toggleRowExpansion(row, false)
-        }
-      }
-
-      dataList.value.forEach(row => {
-        if (expanded.value) {
-          collapseRecursively(row)
-        } else {
-          expandRecursively(row)
-        }
-      })
-    }
-
-    const isLoadingReport = computed(() => {
-      return store.getters.getReportIsLoading
-    })
-    watch(data, () => {
-      nextTick(() => {
-        expandedRowAll()
-      })
-    })
-    watch(expanded, () => {
-      nextTick(() => {
-        expandedRowAll()
-      })
-    })
-    onMounted(() => {
-      nextTick(() => {
-        expandedRowAll()
-      })
-    })
-    function widthColumn() {
-      dataList.value.forEach(data => {
-        Object.values(data.cells).forEach(dataCell => {
-          return dataCell.display_value.length
-        })
-      })
-    }
-    function getCellStyle(code, row) {
-      if (isEmptyValue(row.cells[code])) {
-        return {}
-      }
-      const { value } = row.cells[code]
-      if (!isEmptyValue(value) && value.type) {
-        if (value.type === 'decimal' && value.value < 0) {
-          return { color: 'red' }
-        }
-      }
-    }
-
-    function shouldHideName(row, code, columnIndex) {
-      if (columnIndex !== 0) {
-        return false
-      }
-      if (row.level > 1 && row.is_parent) {
-        return row.children.some(child => displayLabel(code, child) === displayLabel(code, row))
-      }
-      return false
-    }
     return {
+      // Computed
       showDialog,
-      tableReportEngine,
-      selectedRow,
-      selectedColumn,
-      showPopover,
-      dataModal,
-      dataList,
-      recordData,
-      currentPageSize,
-      currentPageNumber,
-      expanded,
-      isLoadingReport,
-      data,
-      columns,
       shortsKey,
-      reportDefinition,
-      height,
+      // Methods
       keyAction,
-      widthColumn,
-      exportFile,
-      displayLabel,
-      getAlignment,
-      tableRowClassName,
-      handleChangeSizePage,
-      handleChangePage,
-      handleRowClick,
-      getRowClassName,
-      findParent,
-      expandedRowAll,
-      viewShowDialog,
-      getCellStyle,
-      activatePopover,
-      shouldHideName
+      viewShowDialog
     }
   }
 })
@@ -438,56 +135,4 @@ export default defineComponent({
       text-align: right;
     }
   }
-</style>
-<style>
-.containerReportEnginer .el-card__body {
-  padding: 20px !important
-}
-:root {
-  --level-offset: 20px;
-}
-.el-table__row--level-[n] {
-  transform: translateX(calc(var(--level-offset) * var(--level, 1)));
-}
-.expanded {
-  border-left: 3px solid #ddd;
-  height: 50%;
-  position: absolute;
-  left: 0;
-  top: 0;
-  border-bottom: 5px solid transparent;
-  border-top: 5px solid #ddd;
-  width: 0;
-  height: 0;
-  position: absolute;
-  left: -5px;
-}
-
-.arrow-right {
-  position: absolute;
-  bottom: -6px;
-  left: 100%;
-  margin-left: 5px;
-  width: 5px;
-  height: 5px;
-  border-top: 5px solid transparent;
-  border-bottom: 5px solid transparent;
-  border-left: 8px solid #ddd;
-  content: '';
-}
-.last-child-row {
-    border-bottom: 1px solid #afaeae !important;
-}
-.el-table .success-row {
-  font-weight: 700;
-  background: #ecf5ff;
-}
-.reportInfo .el-popover {
-  width: 850px !important;
-}
-.reportInfo .el-popover .el-descriptions-item__container .el-descriptions-item__content{
-  display: flex;
-  justify-content: flex-end;
-  margin-right:20px;
-}
 </style>
